backend/pyproject.toml
[project]
name = "business-template-backend"
version = "0.1.0"
description = "Generalized business back-office template — backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "passlib[argon2]>=1.7",
    "python-multipart>=0.0.20",
    "aiosmtplib>=3.0",
    "jinja2>=3.1",
    "redis>=5.2",
    "slowapi>=0.1.9",
    "boto3>=1.35",
    "httpx>=0.28",
    "pillow>=11.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W", "C4", "SIM", "ASYNC"]
# E501 (line too long) is disabled: `ruff format` wraps lines at line-length
# automatically, so we don't want the linter to complain about cases it can't
# safely reformat (e.g., long URLs in strings).
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"alembic/**" = ["E402", "I001", "UP007", "UP035"]
"tests/**" = ["E402", "B017", "SIM117"]
"app/modules/*/router.py" = ["B008"]
"app/core/auth.py" = ["B008"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
